rich [ci skip]
We are dropping setting of the width of tables columns, previously
fixed, and now let rich adjust it based on content. This might be
changed back later on, when we'll get a better idea of what we want.

We drop the screen() view function and call header(), processes() and
footer() directly in ui.main(). When calling header(), we then compute
the number of lines on screen and pass on the value to processes() in
order to keep scrolling working. Accordingly, the ActivityStats type
alias is dropped as it's sole purpose was to being passed as an argument
to screen().

In 'colors' module, we rename all color "code names" to remove
underscores so that these can be used as rich's style.

Function activities.sorted(), now used directly in ui.main(), has its
type changed to match actual usage (more general type variable).

TODO:
* update all tests
* port the overlay window to cancel/terminate a process
* remove blessed completely